Set Gtk.Builder ui files translation
*Why we should replace gettext with locale*

Thanks to Juha Sahakangas on the #gtk+ IRC channel for providing the explanation: For this particular case the locale module needs to be used instead of gettext. Python's gettext module is pure python, it doesn't actually set the text domain in a way that the C library can read, but locale does (by calling libc). So long as that's done, GtkBuilder already works the way you're asking it to.

*from https://stackoverflow.com/questions/24320502/how-to-translate-pygtk-glade-gtk-builder-application *